+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Add ⚡ to the list of supported Bitcoin Lightning labels.
 
+### Fixed
+
+- Remove "Follow" button from user's own profile page.
+
 ## [2.7.0] - 2024-01-12
 
 ### Added